Refactor: Moves to nested api routes
Also adds the ability to request only a certain value of records when
requesting all price records
