+++ next-app/src/pages/api/item/get-all-prices.ts
@@ -2,7 +2,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { z } from 'zod'
 
-import { getAllItemsPrices } from '../../lib/db'
+import { getAllItemsPrices } from '../../../lib/db'
 
 // type Data = {
 //   name: string
@@ -23,7 +23,8 @@ export default async function handler(
   const validatedRequestedItem = requestedItem.safeParse(item)
 
   if (validatedRequestedItem.success) {
-    const itemPrices = await getAllItemsPrices(validatedRequestedItem.data.id)
+    const numberOfRecords = validatedRequestedItem.data.numberOfPriceRecords ?? 30
+    const itemPrices = await getAllItemsPrices(validatedRequestedItem.data.id, numberOfRecords)
 
     res.status(200).json({
       item: itemPrices,
