+++ next-app/src/pages/api/item/get-items-and-latest-prices.ts
@@ -1,7 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { PriceRecord, Item } from '@prisma/client';
 import type { NextApiRequest, NextApiResponse } from 'next'
-import { getLatestItemPrices } from '../../lib/db';
+import { getLatestItemPrices } from '../../../lib/db';
 
 type Data = Item & { prices: PriceRecord[]}
